Add files via upload
In this GitHub repository, you will find a hospital database designed and developed using Microsoft SQL Server Management Studio. The repository contains multiple resources, including a database tables Excel file, a .bak file for backup and restoration, an entity-relationship diagram (ERD) to visualize the database schema, and a Balsamiq page for the graphical user interface design.

The Excel file contains all the tables used in the database, including the relationships between them. The .bak file can be used to easily restore the database in case of data loss or corruption. The ERD diagram visually represents the relationships between the different tables in the database. The Balsamiq page provides an interactive visual representation of the user interface design for the hospital database.